ddsc_web/ddsc_web/settings/custom_storages.py
from django.conf import settings
from storages.backends.s3boto3 import S3Boto3Storage, S3StaticStorage


class StaticStorage(S3StaticStorage):
    location = settings.STATICFILES_LOCATION
    default_acl = "public-read"

    def get_object_parameters(self, name):
        return {
            "CacheControl": "public, max-age=86400",
        }


class PublicMediaStorage(S3Boto3Storage):
    location = settings.PUBLIC_MEDIA_LOCATION
    file_overwrite = False
    default_acl = "public-read"


class PrivateMediaStorage(S3Boto3Storage):
    location = settings.PRIVATE_MEDIA_LOCATION
    default_acl = "private"
    file_overwrite = False
    custom_domain = False
