requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
black==22.12.0
bleach==5.0.1
boto3==1.26.50
botocore==1.29.50
cachetools==5.2.1
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
diff-match-patch==20230430
Django==4.1.5
django-appconf==1.0.5
django-crispy-forms==1.14.0
django-image-cropping==1.7
django-import-export==4.0.3
django-markdownfield==0.10.0
django-storages==1.13.2
django-tinymce==3.5.0
easy-thumbnails==2.8.5
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
jmespath==1.0.1
kombu==5.2.4
Markdown==3.4.1
mypy-extensions==0.4.3
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
protobuf==4.21.12
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
qrcode==7.3.1
redis==4.4.2
requests==2.28.2
rsa==4.9
s3transfer==0.6.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
tablib==3.5.0
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
